Fix imports so that common has no portal dependencies
