+++ cfl_common/common/email_messages.py
