+++ portal/tests/test_migration_use_common_models.py
@@ -18,3 +18,17 @@ class TestMigrationUseCommonModels(MigrationTestCase):
         for moved_model in moved_models:
             assert f"portal_{moved_model}" not in model_names
             assert f"common_{moved_model}" in model_names
+
+
+class TestMigrationMoveEmailVerificationToCommon(MigrationTestCase):
+
+    start_migration = "0058_move_to_common_models"
+    dest_migration = "0059_move_email_verifications_to_common"
+
+    def test_emailverification_moved_to_common(self):
+        model_names = [
+            model._meta.db_table for model in self.django_application.get_models()
+        ]
+
+        assert f"portal_emailverification" not in model_names
+        assert f"common_emailverification" in model_names
