+++ portal/views/student/play.py
@@ -41,7 +41,7 @@ from django.core.urlresolvers import reverse_lazy
 from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render
 
-from portal import email_messages
+from common import email_messages
 from portal.forms.play import StudentJoinOrganisationForm
 from common.helpers.emails import send_email, NOTIFICATION_EMAIL
 from common.permissions import logged_in_as_student, logged_in_as_independent_student
